requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.3
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
fontawesomefree==6.2.1
mysqlclient==2.1.1
pilkit==2.0
Pillow==9.3.0
pytz==2022.6
sqlparse==0.4.3
